Updated e2e test
